+++ src/app/settings/views/Users/UserDelete/UserDelete.test.tsx
@@ -1,3 +1,5 @@
+import configureStore from "redux-mock-store";
+
 import UserDelete from "./UserDelete";
 
 import type { RootState } from "@/app/store/root/types";
@@ -10,6 +12,8 @@ import {
 import { renderWithBrowserRouter, screen } from "@/testing/utils";
 
 let state: RootState;
+
+const mockStore = configureStore<RootState>();
 
 beforeEach(() => {
   state = rootStateFactory({
@@ -40,4 +44,17 @@ it("renders", () => {
     routePattern: "/settings/users/:id/edit",
   });
   expect(screen.getByRole("form", { name: "Delete user" }));
+});
+
+it("can add a message when a user is deleted", () => {
+  state.user.saved = true;
+  const store = mockStore(state);
+  renderWithBrowserRouter(<UserDelete />, {
+    store,
+    route: "/settings/users/1/edit",
+    routePattern: "/settings/users/:id/edit",
+  });
+  const actions = store.getActions();
+  expect(actions.some((action) => action.type === "user/cleanup")).toBe(true);
+  expect(actions.some((action) => action.type === "message/add")).toBe(true);
 });
